# api/openapi.yaml
openapi: 3.1.0
info:
  title: KUVTH Zebrafish LIMS API
  version: 1.0.0
  summary: Zebrafish somatic cell nuclear transfer embryo and clone-fish tracking
  description: |
    Contract between the static SPA front end and the Python back end.

    This document is the single source of truth for the interface
    (SRS v1.0 section 8, API-02). Generate both sides from it:

      * server stubs — generated directly from this OpenAPI 3.1 document
      * browser client — `openapi-typescript api/openapi.yaml -o frontend/src/api/schema.d.ts`

    ### Conventions

    * All timestamps are RFC 3339 **with an explicit offset** (CI-04).
      The server stores UTC; the client sends and displays `Asia/Bangkok`.
    * All hour values are decimal hours with 4 decimal places (DR-05).
    * All API routes except health and the two OTP entry routes require an
      authenticated session cookie. Membership is invitation-only and limited
      to `@ku.th` addresses; authorization is enforced by the server.
    * Experimental writes carry `X-Operator-Id`, `X-Device-Id`, and the required
      `X-Actor-User-Id`. The account actor comes from the session and must
      match this header so queued writes cannot replay under another account.
      Correction request writes use the account, device and idempotency key;
      they do not require a linked experimental operator and must be online.
    * Every mutation also accepts `X-Idempotency-Key`. Clients persist this
...
      queue retries freely.
...
servers:
...
security:
  - sessionCookie: []
...
paths:
  /auth/request-code:
    post:
      security: []
      tags: [Authentication]
      summary: Request a single-use email code
      operationId: requestSignInCode
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestCodeInput' }
      responses:
        '202': { description: "Same response for invited, unknown, and rate-limited addresses; delivery is queued" }
        '429': { description: Global per-IP API rate limit reached }
        '503': { description: Email delivery is not configured }
  /auth/verify-code:
    post:
      security: []
      tags: [Authentication]
      summary: Verify a code and create a persistent browser session
      operationId: verifySignInCode
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyCodeInput' }
      responses:
        '200':
          description: Session created; token is returned only as an HttpOnly cookie
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/AuthUser' }
        '401': { description: Invalid, expired, or already-used code }
...
  /timing-profiles/csv:
    get:
      tags: [Timing]
      summary: Download the current reference timing as CSV
      operationId: exportTimingCsv
      parameters:
        - name: protocolId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: CSV with columns stage_order, stage_code, label, expected_hpa
          content:
            text/csv:
              schema: { type: string }
    post:
      tags: [Timing]
      summary: Upload reference timing as CSV, creating a new version
      operationId: importTimingCsv
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey', $ref: '#/components/parameters/ActorUserId' ]
      requestBody:
        required: true
        content:
          text/csv:
            schema: { type: string }
      responses:
        '201':
          description: New version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimingProfile' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

  # --------------------------------------------------------------- experiments
...
  /observations/fish:
    post:
      tags: [Observations]
      summary: Record fish observations (bulk, idempotent)
      description: |
        Same idempotency contract as the embryo endpoint (BR-18). "Everyone still
        alive" is one call carrying every living fish (FR-604, AC-604).

        An outcome of `DEAD`, `FROZEN`, or `DISCARDED` also closes the fish:
        `status`, `exitDate`, and `exitReason` are set, and it drops off the next
        day's roll-call (FR-606, AC-606).

        For a consecutive date range, the client expands each fish/date into a
        separate item with its own `clientUuid`. This keeps retries idempotent.
      operationId: createFishObservations
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey', $ref: '#/components/parameters/ActorUserId' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [observations]
              properties:
                observations:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: { $ref: '#/components/schemas/FishObservationInput' }
      responses:
        '200':
          description: Per-item results
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/FishObservationResult' }
  /observations/fish/{id}:
    parameters: [ $ref: '#/components/parameters/PathId' ]
    patch:
      tags: [Observations]
      summary: Correct a fish observation
      operationId: updateFishObservation
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey', $ref: '#/components/parameters/ActorUserId' ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FishObservationPatchInput' }
      responses:
        '204': { description: Corrected and audited }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      tags: [Observations]
      summary: Soft-delete an incorrect fish observation
      operationId: deleteFishObservation
      parameters:
        - $ref: '#/components/parameters/OperatorId'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorUserId'
        - name: reason
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 500 }
      responses:
        '204': { description: Soft-deleted and audited }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------- promotion
...
  /fish:
    get:
      tags: [Fish]
      summary: The clone fish register
      operationId: listFish
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/FishStatus' } }
        - { name: siteId, in: query, schema: { type: string, format: uuid } }
        - { name: boxId, in: query, schema: { type: string, format: uuid } }
        - { name: experimentGroupId, in: query, schema: { type: string, format: uuid } }
        - { name: treatmentGroupId, in: query, schema: { type: string, format: uuid } }
        - { name: batchId, in: query, schema: { type: string, format: uuid } }
        - { name: operatorId, in: query, schema: { type: string, format: uuid } }
        - { name: dateFrom, in: query, schema: { type: string, format: date } }
        - { name: dateTo, in: query, schema: { type: string, format: date } }
        - { name: donorCellLineId, in: query, schema: { type: string, format: uuid } }
        - { name: strain, in: query, schema: { type: string } }
        - { name: condition, in: query, schema: { $ref: '#/components/schemas/Condition' } }
        - { name: dobFrom, in: query, schema: { type: string, format: date } }
        - { name: dobTo, in: query, schema: { type: string, format: date } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Fish
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CloneFish' }
                  nextCursor: { type: [string, 'null'] }
    post:
      tags: [Fish]
      summary: Register a fish by hand
      description: For fish that did not come through the normal promotion path (FR-509).
      operationId: createFish
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey', $ref: '#/components/parameters/ActorUserId' ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CloneFishInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CloneFish' }
        '409': { $ref: '#/components/responses/Conflict' }
  /fish/{id}:
    parameters: [ $ref: '#/components/parameters/PathId' ]
    get:
      tags: [Fish]
      summary: One fish with its full observation timeline
      operationId: getFish
      responses:
        '200':
          description: Fish detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CloneFishDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Fish]
      summary: Update a fish
      description: Sex is usually only determinable weeks in (FR-609), so it can be set at any time.
      operationId: updateFish
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey', $ref: '#/components/parameters/ActorUserId' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fishCode: { type: string }
                sex: { $ref: '#/components/schemas/Sex' }
                finClipped: { type: boolean }
                fishBoxId: { type: [string, 'null'], format: uuid }
                remarks: { type: [string, 'null'] }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CloneFish' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /audit-log:
    get:
      tags: [Audit]
      summary: Admin-only change history
      operationId: listAuditLog
      parameters:
        - { name: table, in: query, schema: { type: string, maxLength: 128 } }
        - { name: recordId, in: query, schema: { type: string, maxLength: 36 } }
        - { name: auditId, in: query, schema: { type: string, format: uuid } }
        - { name: operatorId, in: query, schema: { type: string, format: uuid } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEntry' }
                  nextCursor: { type: [string, 'null'] }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: chronofish_session
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    OperatorId:
      name: X-Operator-Id
      in: header
      required: true
      description: Experimental operator associated with the recorded work.
      schema: { type: string, format: uuid }
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      description: Stable per-device identifier generated on first use and kept in local storage.
      schema: { type: string, maxLength: 64 }
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      description: Stable key for one logical mutation. Replays return the original result.
      schema: { type: string, format: uuid }
    ActorUserId:
      name: X-Actor-User-Id
      in: header
      required: true
      description: Account id for this write; must match the authenticated session.
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 512 }
...
  responses:
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflicts with current state — for example a duplicate normalised name, or a monotonic-survival violation without an override reason
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Well-formed but breaks a business rule (BR-19)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            beforeActivation:
              summary: Observation timed before its lot was activated
              value:
                error:
                  code: OBSERVATION_BEFORE_ACTIVATION
                  message: เวลาที่สังเกตอยู่ก่อนเวลา activation ของล็อตนี้
                  field: observedAt
                  details:
                    observedAt: '2026-08-20T09:00:00+07:00'
                    activatedAt: '2026-08-20T10:59:00+07:00'
  schemas:
...
    FishOutcome:
      type: string
      enum: [ALIVE, DEAD, FROZEN, DISCARDED, NOT_OBSERVED]
    Condition:
      type: string
      enum: [NORMAL, ABNORMAL, UNDETERMINED]
    FishHealthStatus:
      type: string
      enum: [HEALTHY, WEAK, SICK, DISABLED, AGED, UNDETERMINED]
    FishStatus:
      type: string
      enum: [ALIVE, DEAD, FROZEN, DISCARDED]
    Sex:
      type: string
      enum: [M, F, UNKNOWN]
...
    SpecimenKind:
      type: string
      enum: [CL, RT, DC]
      description: Clone, Recipient Tail, Donor Cell.
    SpecimenType:
      type: string
      enum: [WHOLE_EMBRYO, CAUDAL_FIN_CLIP]
    WriteStatus:
      type: string
      enum: [created, duplicate, rejected]
      description: '`duplicate` is a normal outcome of retrying, not a failure.'

    # ---- errors -------------------------------------------------------------
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              examples: [OBSERVATION_BEFORE_ACTIVATION]
            message:
              type: string
              description: Written for the person on the iPad, not for a log file (NFR-304).
            field: { type: [string, 'null'] }
            details: { type: [object, 'null'], additionalProperties: true }

    # ---- master data --------------------------------------------------------
...
    TimingProfile:
      type: object
      required: [id, protocolId, version, name, isCurrent, entries]
      properties:
        id: { type: string, format: uuid }
        protocolId: { type: string, format: uuid }
        version: { type: integer, examples: [1] }
        name: { type: string, examples: ['ZFIN 28.5C (default)'] }
        referenceTempC: { type: [number, 'null'], examples: [28.5] }
        autoTempAdjust:
          type: boolean
          description: Always false in v1 (Q-N2); the field exists so enabling FR-207 later needs no migration.
        sourceNote: { type: [string, 'null'], examples: ['Kimmel et al. 1995 / ZFIN, 28.5 C'] }
        isCurrent: { type: boolean }
        createdAt: { type: string, format: date-time }
        createdByOperatorId: { type: [string, 'null'], format: uuid }
        entries:
          type: array
          items:
            type: object
            required: [stageCode, expectedHpa]
            properties:
              stageOrder: { type: integer, description: Derived from stageCode when omitted. }
              stageCode: { type: string }
              stageLabel: { type: string, description: Derived from the canonical stage when omitted. }
              expectedHpa: { type: number, examples: [2.5] }

    # ---- experiments --------------------------------------------------------
...
    FishObservationInput:
      type: object
      required: [clientUuid, cloneFishId, observedOn, outcome, condition, healthStatus]
      properties:
        clientUuid: { type: string, format: uuid }
        cloneFishId: { type: string, format: uuid }
        observedOn: { type: string, format: date }
        outcome: { $ref: '#/components/schemas/FishOutcome' }
        condition: { $ref: '#/components/schemas/Condition' }
        healthStatus: { $ref: '#/components/schemas/FishHealthStatus' }
        notes: { type: [string, 'null'] }
        overrideReason:
          type: [string, 'null']
          maxLength: 500
          description: Required by the server when the observation is backdated (BR-07, BR-19).
    FishObservationPatchInput:
      type: object
      minProperties: 2
      required: [overrideReason]
      properties:
        observedOn: { type: string, format: date }
        outcome: { $ref: '#/components/schemas/FishOutcome' }
        condition: { $ref: '#/components/schemas/Condition' }
        healthStatus: { $ref: '#/components/schemas/FishHealthStatus' }
        notes: { type: [string, 'null'] }
        overrideReason:
          type: string
          minLength: 1
          maxLength: 500
          description: Audit reason for correcting the observation.
    FishObservationResult:
      type: object
      required: [clientUuid, status]
      properties:
        clientUuid: { type: string, format: uuid }
        id: { type: [string, 'null'], format: uuid }
        status: { $ref: '#/components/schemas/WriteStatus' }
        ageDays: { type: [integer, 'null'] }
        fishClosed:
          type: [boolean, 'null']
          description: True when this observation ended the fish's tracking (FR-606).
        error: { type: [object, 'null'], additionalProperties: true }

    # ---- fish ---------------------------------------------------------------
...
    CloneFish:
      type: object
      required: [id, fishCode, runningNo, dob, donorCellLineId, status, condition, sex, finClipped]
      properties:
        id: { type: string, format: uuid }
        embryoId: { type: [string, 'null'], format: uuid }
        embryoCode: { type: [string, 'null'] }
        fishCode: { type: string, examples: ['No.6_Clone3-AB cell-16'] }
        runningNo: { type: integer }
        dob: { type: string, format: date }
        ageDays:
          type: integer
          description: Computed live from dob, never stored (BR-11, DC-04).
        strain: { type: string }
        donorCellLineId: { type: string, format: uuid }
        recipientEggLotId: { type: [string, 'null'], format: uuid }
        siteId: { type: [string, 'null'], format: uuid }
        fishBoxId: { type: [string, 'null'], format: uuid }
        fishBoxCode: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/FishStatus' }
        condition: { $ref: '#/components/schemas/Condition' }
        healthStatus: { $ref: '#/components/schemas/FishHealthStatus' }
        firstAbnormalOn: { type: [string, 'null'], format: date }
        firstAbnormalAgeDays: { type: [integer, 'null'] }
        sex: { $ref: '#/components/schemas/Sex' }
        finClipped: { type: boolean }
        exitDate: { type: [string, 'null'], format: date }
        exitReason: { type: [string, 'null'], enum: [DEAD, FROZEN, DISCARDED, LOST, null] }
        remarks: { type: [string, 'null'] }
    CloneFishInput:
      type: object
      required: [fishCode, dob, donorCellLineId, recipientEggLotId]
      properties:
        fishCode: { type: string, minLength: 1, maxLength: 150 }
        dob: { type: string, format: date }
        donorCellLineId: { type: string, format: uuid }
        recipientEggLotId: { type: string, format: uuid }
        siteId: { type: [string, 'null'], format: uuid }
        fishBoxId: { type: [string, 'null'], format: uuid }
        condition: { $ref: '#/components/schemas/Condition' }
        healthStatus: { $ref: '#/components/schemas/FishHealthStatus' }
        sex: { $ref: '#/components/schemas/Sex' }
        remarks: { type: [string, 'null'] }
        overrideReason:
          type: [string, 'null']
          maxLength: 500
          description: Required by the server when manually registering an older fish (BR-07).
    CloneFishDetail:
      allOf:
        - $ref: '#/components/schemas/CloneFish'
        - type: object
          properties:
            observations:
              type: array
              items:
                type: object
                required: [id, observedOn, ageDays, outcome, condition]
                properties:
                  id: { type: string, format: uuid }
                  observedOn: { type: string, format: date }
                  ageDays: { type: integer }
                  outcome: { $ref: '#/components/schemas/FishOutcome' }
                  condition: { $ref: '#/components/schemas/Condition' }
                  operatorName: { type: [string, 'null'] }
                  isBackdated: { type: boolean }
                  notes: { type: [string, 'null'] }
            specimens:
              type: array
              items: { $ref: '#/components/schemas/Specimen' }
            embryoTimeline:
              type: array
              description: The Stage 1 history of the embryo this fish came from, so one screen covers the whole life.
              items: { $ref: '#/components/schemas/EmbryoObservation' }
    Specimen:
      type: object
      required: [id, cloneFishId, specimenCode, specimenKind, specimenType]
      properties:
        id: { type: string, format: uuid }
        cloneFishId: { type: string, format: uuid }
        specimenCode: { type: string, examples: [CL1] }
        specimenKind: { $ref: '#/components/schemas/SpecimenKind' }
        specimenType: { $ref: '#/components/schemas/SpecimenType' }
        collectedOn: { type: [string, 'null'], format: date }
        frozenOn: { type: [string, 'null'], format: date }
        storage: { type: [string, 'null'], enum: ['-20', '-80', null] }
        notes: { type: [string, 'null'] }
...
    AuditEntry:
      type: object
      required: [id, tableName, recordId, action, occurredAt]
      properties:
        id: { type: string, format: uuid }
        tableName: { type: string }
        recordId: { type: string, maxLength: 36 }
        action: { type: string, enum: [INSERT, UPDATE, DELETE] }
        oldValues: { type: [object, 'null'], additionalProperties: true }
        newValues: { type: [object, 'null'], additionalProperties: true }
        operatorId: { type: [string, 'null'], format: uuid }
        operatorName: { type: [string, 'null'] }
        deviceId: { type: [string, 'null'] }
        actorUserId: { type: [string, 'null'], format: uuid }
        actorEmail: { type: [string, 'null'], format: email }
        occurredAt: { type: string, format: date-time }
    RequestCodeInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, pattern: '^[^@]+@ku\.th$' }
    VerifyCodeInput:
      type: object
      required: [email, code]
      properties:
        email: { type: string, format: email, pattern: '^[^@]+@ku\.th$' }
        code: { type: string, pattern: '^[0-9]{6}$' }
    AuthUser:
      type: object
      required: [id, email, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { type: string, enum: [admin, member] }
        operatorId: { type: [string, 'null'], format: uuid }
        sessionId: { type: [string, 'null'], format: uuid }
        deviceId: { type: [string, 'null'] }
...